test: add test for os value
