+++ dc_util/package.xml
@@ -16,6 +16,7 @@
   <depend>launch</depend>
   <depend>launch_testing_ament_cmake</depend>
   <depend>nav2_util</depend>
+  <depend>nlohmann-json-dev</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_lifecycle</depend>
   <depend>yaml_cpp_vendor</depend>
